+++ solver.c
@@ -38,7 +38,7 @@ void revuchararr(unsigned char a[8]){
 
 //convert a number into an unsigned char array
 void num2uchararray(long unsigned int v, unsigned char a[8]){
-	int x = v; //get value to work with
+	long unsigned int x = v; //get value to work with
 	int it = 7; //iterator for placing into the array
 	while(x>0 && it>=0){ //while not at max array index and x is nonzero
 		int temp = x%256; //get first 256 remainder
